+++ R/CallableFunction.R
@@ -7,6 +7,7 @@
 #' @importFrom rlang parse_expr
 CallableFunction <- R6::R6Class( #nolint
   "CallableFunction",
+  inherit = Callable,
 
   ## __Public Methods ====
   public = list(
@@ -20,34 +21,19 @@ CallableFunction <- R6::R6Class( #nolint
     #'
     #' @return new \code{CallableFunction} object
     initialize = function(fun, env = new.env(parent = parent.env(globalenv()))) {
-      stopifnot(is.environment(env))
+      super$initialize(env = env)
+
+      stop_if_not(
+        list(!missing(fun),
+             "A valid function name must be provided.")
+      )
 
       fun_name <- private$get_callable_function(fun)
       private$fun_name <- pdeparse(fun_name)
-      private$env <- env
-      private$refresh()
-      invisible(self)
-    },
 
-    #' @description
-    #' Assigns \code{x <- value} object to \code{env}
-    #' @param x (\code{character} value)\cr
-    #'  name of the variable in class environment
-    #' @param value (\code{data.frame})\cr
-    #'  object to be assigned to \code{x}
-    #'
-    #' @return arguments the function gets called with
-    assign_to_env = function(x, value) {
-      assign(x, value, envir = private$env)
+      private$refresh()
+
       return(invisible(self))
-    },
-    #' @description
-    #' Check if evaluation of the function has not failed.
-    #'
-    #' @return \code{TRUE} if evaluation of the function failed or \code{FALSE}
-    #'  if evaluation failed or function hasn't yet been called.
-    is_failed = function() {
-      return(private$failed)
     },
     #' @description
     #' get the arguments a function gets called with
@@ -86,52 +72,6 @@ CallableFunction <- R6::R6Class( #nolint
       }
 
       return(res)
-    },
-    #' @description
-    #' Get error message from last function execution
-    #'
-    #' @return \code{try-error} object with error message or \code{character(0)} if last
-    #'  function evaluation was successful.
-    get_error_message = function() {
-      return(private$error_msg)
-    },
-    #' @description
-    #' Run function
-    #'
-    #' @param return (\code{logical} value)\cr
-    #'  whether to return an object
-    #' @param args (\code{NULL} or named \code{list})\cr
-    #'  dynamic arguments passed to function. Dynamic arguments are executed in this call and are not
-    #'  saved which means that \code{self$get_call()} won't include them later.
-    #' @param try (\code{logical} value)\cr
-    #'  whether perform function evaluation inside \code{try} clause
-    #'
-    #' @return nothing or output from function depending on \code{return} argument
-    run = function(return = TRUE, args = NULL, try = FALSE) {
-      stopifnot(is_logical_single(return))
-      stopifnot(is_empty(args) || is_fully_named_list(args))
-      stopifnot(is_logical_single(try))
-
-      # args are "dynamic" are used only to evaluate this call
-      # - args not saved to private$call persistently
-      expr <- self$get_call(deparse = FALSE, args = args)
-
-      res <- if (try) {
-        try(
-          eval(expr, envir = private$env),
-          silent = TRUE
-        )
-      } else {
-        eval(expr, envir = private$env)
-      }
-
-      private$check_run_output(res)
-
-      if (return) {
-        return(res)
-      } else {
-        return(invisible(NULL))
-      }
     },
     #' @description
     #' Set up function arguments
@@ -183,23 +123,7 @@ CallableFunction <- R6::R6Class( #nolint
   private = list(
     fun_name = character(0),
     args = NULL, # named list with argument names and values
-    call = NULL, # a call object
-    env = NULL, # environment where function is called
-    failed = FALSE,
-    error_msg = character(0),
     ## __Private Methods ====
-    # check if the function was evaluated properly - if not keep error message
-    check_run_output = function(res) {
-      if (is(res, "try-error")) {
-        private$failed <- TRUE
-        private$error_msg <- res
-      } else {
-        private$failed <- FALSE
-        private$error_msg <- character(0)
-      }
-
-      return(NULL)
-    },
     # @description
     # Refresh call with function name and saved arguments
     #