+++ R/get_dataset_label.R
@@ -19,14 +19,14 @@ get_dataset_label <- function(x) {
 #' fun$set_args(list(n = 5, seed = 1, cached = TRUE))
 #'
 #' x <- named_dataset_connector(
-#'  pull_fun = fun,
+#'  pull_callable = fun,
 #'  dataname = "ADSL"
 #' )
 #' get_dataset_label(x)
 #'
 #' x1 <- relational_dataset_connector(
 #'  dataname = "ADSL",
-#'  pull_fun = fun,
+#'  pull_callable = fun,
 #'  keys = get_cdisc_keys("ADSL")
 #' )
 #' get_dataset_label(x1)