+++ R/is_pulled.R
@@ -15,7 +15,7 @@ is_pulled <- function(x) {
 #'
 #' @examples
 #' # RawDatasetConnector --------
-#' ds <- raw_dataset_connector(pull_fun = callable_function(data.frame))
+#' ds <- raw_dataset_connector(pull_callable = callable_function(data.frame))
 #' set_args(ds, list(x = 1:5, y = letters[1:5], stringsAsFactors = FALSE))
 #'
 #' is_pulled(ds)